[Loom] Factor source memory planning by responsibility
Move dependency and scheduling summary derivation out of canonical
source memory plan construction without introducing a private cross-file
API.

Split the oversized test source around a shared fixture while preserving
a single test binary and suite. Fixture failures now use the gtest
status matchers instead of terminating the process through
IREE_CHECK_OK.
