+++ loom/src/loom/codegen/low/BUILD.bazel
@@ -728,7 +728,10 @@ loom_cc_library(
 
 loom_cc_library(
     name = "source_memory_plan",
-    srcs = ["source_memory_plan.c"],
+    srcs = [
+        "source_memory_plan.c",
+        "source_memory_plan_summary.c",
+    ],
     hdrs = ["source_memory_plan.h"],
     deps = [
         ":memory_access",
@@ -764,7 +767,11 @@ loom_cc_benchmark(
 
 loom_cc_test(
     name = "source_memory_plan_test",
-    srcs = ["source_memory_plan_test.cc"],
+    srcs = [
+        "source_memory_plan_index_test.cc",
+        "source_memory_plan_test.cc",
+        "source_memory_plan_test_fixture.h",
+    ],
     deps = [
         ":source_memory_plan",
         "//loom/src/loom/analysis:view_regions",
